fix(router): Truncate overlong error message
This fixes a panic that could be caused when the router tried to send
an error message through the FFI that was longer than the static limit
of 2048 characters.  This is now fixed so that if the message is
longer than the static limit, it is truncated.  The limit has also
been increased to 4096.

Fixes KAG-458