KAG-458: ATC router limits Route.paths to 2048 bytes (#27)
* tests(*): add test case for errbuf length

* fix(compilation): Allow deprecated features in parser to avoid warnings

* fix(router): Truncate overlong error message

This fixes a panic that could be caused when the router tried to send
an error message through the FFI that was longer than the static limit
of 2048 characters.  This is now fixed so that if the message is
longer than the static limit, it is truncated.  The limit has also
been increased to 4096.

Fixes KAG-458

* feat(ci): Run unit tests during build

* fix(ffi): set size pointer to actual buffer size

* fix(*): remove debug output

Co-authored-by: Michael Martin <3277009+flrgh@users.noreply.github.com>