feat: derive `PartialEq`, `Eq` for `ast::Expression` (#315)
Strictly speaking, AST is not `PartialEq` and `Eq`, this because both are requires symmetry property.

In short, if AST is symmetric, then `a == b` if and only if `b == a`, however, clearly, AST is not symmetric.
However, [`syn`] derives the `PartialEq` and `Eq` for AST types like [`syn::ExprAsync`].

Since the popular crate has already did it, it should be ok for us to do the similar thing.

[`syn`]: https://docs.rs/syn/latest/syn
[`syn::ExprAsync`]: https://docs.rs/syn/2.0.104/syn/struct.ExprAsync.html#impl-PartialEq-for-ExprAsync