feat(*): schema owned or referenced
This commit introduces a new way of initializing ATC so that it owns the
schema that it depends on, instead of relying on a shared reference.

This facilitates the workflow especially when ATC is instantiated and
used in Rust libraries, particularly in use cases where it needs to be
cached for reuse - so that it is not necessary to maintain a reference
to an external Schema object.

This is based on @lucab 's work and includes some small changes that are
a follow up to his observations (same PR).

HOUD-98

Co-authored-by: Luca BRUNO <lucab@lucabruno.net>